Implemented guards against graph cycles and unit tests for same. Fixed typos and other linter problems.
